Revert "Merge pull request #3615 from cphyc/remove-duplicate-code"
This reverts commit 7704cd01d59ba3f318dda492a72ff8817733952b, reversing
changes made to b7dfe4a1676c7267f94b09753bfb80b9875f9894.
