+++ yt/frontends/chombo/data_structures.py
@@ -8,7 +8,6 @@ import numpy as np
 from yt.data_objects.index_subobjects.grid_patch import AMRGridPatch
 from yt.data_objects.static_output import Dataset
 from yt.fields.field_info_container import FieldInfoContainer
-from yt.frontends.boxlib.misc import BoxlibReadParticleFileMixin
 from yt.funcs import mylog, setdefaultattr
 from yt.geometry.grid_geometry_handler import GridIndex
 from yt.utilities.file_handler import HDF5FileHandler, warn_h5py
@@ -578,7 +577,7 @@ class PlutoDataset(ChomboDataset):
         return False
 
 
-class Orion2Hierarchy(ChomboHierarchy, BoxlibReadParticleFileMixin):
+class Orion2Hierarchy(ChomboHierarchy):
     def __init__(self, ds, dataset_type="orion_chombo_native"):
         ChomboHierarchy.__init__(self, ds, dataset_type)
 
@@ -599,7 +598,46 @@ class Orion2Hierarchy(ChomboHierarchy, BoxlibReadParticleFileMixin):
         self.field_list.extend(pfield_list)
 
     def _read_particles(self):
-        return self._read_particle_file(self.particle_filename)
+        if not os.path.exists(self.particle_filename):
+            return
+        with open(self.particle_filename) as f:
+            lines = f.readlines()
+            self.num_stars = int(lines[0].strip().split(" ")[0])
+            for num, line in enumerate(lines[1:]):
+                particle_position_x = float(line.split(" ")[1])
+                particle_position_y = float(line.split(" ")[2])
+                particle_position_z = float(line.split(" ")[3])
+                coord = [particle_position_x, particle_position_y, particle_position_z]
+                # for each particle, determine which grids contain it
+                # copied from object_finding_mixin.py
+                mask = np.ones(self.num_grids)
+                for i in range(len(coord)):
+                    np.choose(
+                        np.greater(self.grid_left_edge.d[:, i], coord[i]),
+                        (mask, 0),
+                        mask,
+                    )
+                    np.choose(
+                        np.greater(self.grid_right_edge.d[:, i], coord[i]),
+                        (0, mask),
+                        mask,
+                    )
+                ind = np.where(mask == 1)
+                selected_grids = self.grids[ind]
+                # in orion, particles always live on the finest level.
+                # so, we want to assign the particle to the finest of
+                # the grids we just found
+                if len(selected_grids) != 0:
+                    grid = sorted(selected_grids, key=lambda grid: grid.Level)[-1]
+                    ind = np.where(self.grids == grid)[0][0]
+                    self.grid_particle_count[ind] += 1
+                    self.grids[ind].NumberOfParticles += 1
+
+                    # store the position in the *.sink file for fast access.
+                    try:
+                        self.grids[ind]._particle_line_numbers.append(num + 1)
+                    except AttributeError:
+                        self.grids[ind]._particle_line_numbers = [num + 1]
 
 
 class Orion2Dataset(ChomboDataset):
